+++ foqos/Components/BlockedProfileCards/ProfileTimerButton.swift
@@ -1,0 +1,89 @@
+import SwiftUI
+
+struct ProfileTimerButton: View {
+    let isActive: Bool
+    let elapsedTime: TimeInterval?
+    let onStartTapped: () -> Void
+    let onStopTapped: () -> Void
+    
+    var body: some View {
+        HStack(spacing: 8) {
+            if isActive, let elapsedTimeVal = elapsedTime {
+                // Timer with clock icon
+                HStack(spacing: 8) {
+                    Image(systemName: "clock.fill")
+                        .font(.system(size: 14))
+                        .foregroundColor(.primary.opacity(0.7))
+                    
+                    Text(timeString(from: elapsedTimeVal))
+                        .foregroundColor(.primary)
+                        .font(.system(size: 16, weight: .semibold))
+                        .contentTransition(.numericText())
+                        .animation(.default, value: elapsedTimeVal)
+                }
+                .padding(.vertical, 10)
+                .padding(.horizontal, 12)
+                .frame(minWidth: 0, maxWidth: .infinity)
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(.thinMaterial)
+                        .overlay(
+                            RoundedRectangle(cornerRadius: 16)
+                                .stroke(
+                                    Color.primary.opacity(0.2),
+                                    lineWidth: 1
+                                )
+                        )
+                )
+                
+                // Stop button
+                GlassButton(
+                    title: "Stop",
+                    icon: "stop.fill",
+                    fullWidth: false,
+                    equalWidth: true
+                ) {
+                    onStopTapped()
+                }
+            } else {
+                // Start button (full width when no timer is shown)
+                GlassButton(
+                    title: "Hold to Start",
+                    icon: "play.fill",
+                    fullWidth: true,
+                    longPressEnabled: true
+                ) {
+                    onStartTapped()
+                }
+            }
+        }
+    }
+    
+    // Format TimeInterval to HH:MM:SS
+    private func timeString(from timeInterval: TimeInterval) -> String {
+        let hours = Int(timeInterval) / 3600
+        let minutes = Int(timeInterval) / 60 % 60
+        let seconds = Int(timeInterval) % 60
+        return String(format: "%02d:%02d:%02d", hours, minutes, seconds)
+    }
+}
+
+#Preview {
+    VStack(spacing: 20) {
+        ProfileTimerButton(
+            isActive: false,
+            elapsedTime: nil,
+            onStartTapped: {},
+            onStopTapped: {}
+        )
+        
+        ProfileTimerButton(
+            isActive: true,
+            elapsedTime: 3665,
+            onStartTapped: {},
+            onStopTapped: {}
+        )
+    }
+    .padding()
+    .background(Color(.systemGroupedBackground))
+}
